# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: rdr API
  version: 1.0.0
  description: |
    JSON HTTP API for [rdr](https://gitea.nas.lab.chris-allen-lane.com/chrisallenlane/rdr),
    a self-hosted RSS reader. All endpoints under `/api/v1/` (except
    `/api/v1/healthz`) require a bearer token in the `Authorization`
    header.

    Errors follow [RFC 7807 Problem Details](https://tools.ietf.org/html/rfc7807).
    Paginated lists use `Link` (RFC 5988) and `X-Total-Count` headers.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /
    description: Same-origin
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/me:
    get:
      operationId: getMe
      summary: Identify the authenticated user
      description: |
        Returns the user identity associated with the bearer token.
        Useful for clients to confirm their token works and to discover
        their own user id.
      tags: [identity]
      responses:
        "200":
          description: The authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/feeds:
    get:
      operationId: listFeeds
      summary: List feeds
      description: |
        Returns every feed belonging to the authenticated user, ordered
        by title. Pagination is not currently applied — single-user feed
        counts are tractable.
      tags: [feeds]
      responses:
        "200":
          description: Feeds.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Feed"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: addFeed
      summary: Add a feed
      description: |
        Adds a feed by URL. The URL may point directly to a feed or to
        an HTML page; in the latter case the server attempts auto-discovery
        of the feed link. Returns 201 with the newly-created feed as
        soon as the row is inserted; the initial fetch happens in the
        background and may not have populated `title` / `site_url` yet
        when the response is sent.
      tags: [feeds]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFeedRequest"
      responses:
        "201":
          description: Feed created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/feeds/{id}:
    parameters:
      - $ref: "#/components/parameters/IDPath"
    get:
      operationId: getFeed
      summary: Get a feed
      description: |
        Returns a single feed with computed item and unread counts.
        404 if the feed does not exist or does not belong to the caller.
      tags: [feeds]
      responses:
        "200":
          description: The feed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteFeed
      summary: Delete a feed
      description: |
        Removes a feed and cascade-deletes its items.
      tags: [feeds]
      responses:
        "204":
          description: Feed removed.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/feeds/sync:
    post:
      operationId: syncFeeds
      summary: Trigger a feed sync
      description: |
        Asks the background poller to run a sync immediately. Returns
        202 whether or not a sync was actually started (a sync that is
        already in progress is left running rather than queued).
      tags: [feeds]
      responses:
        "202":
          description: Sync trigger accepted.
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/feeds/sync/status:
    get:
      operationId: getSyncStatus
      summary: Get sync status
      description: Returns whether a feed sync is currently in progress.
      tags: [feeds]
      responses:
        "200":
          description: Sync status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/items:
    get:
      operationId: listItems
      summary: List items
      description: |
        Returns a paginated list of items belonging to the authenticated
        user, optionally filtered. Pagination metadata is in the `Link`
        and `X-Total-Count` response headers.
      tags: [items]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - in: query
          name: feed_id
          schema: { type: integer, format: int64, minimum: 1 }
          description: Restrict to a single feed (must belong to the caller).
        - in: query
          name: list_id
          schema: { type: integer, format: int64, minimum: 1 }
          description: Restrict to feeds in a specific list.
        - in: query
          name: unread
          schema: { type: boolean }
          description: If true, only return unread items.
        - in: query
          name: starred
          schema: { type: boolean }
          description: If true, only return starred items.
      responses:
        "200":
          description: Items page. See Link header for pagination.
          headers:
            Link:
              $ref: "#/components/headers/Link"
            X-Total-Count:
              $ref: "#/components/headers/XTotalCount"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/items/{id}:
    parameters:
      - $ref: "#/components/parameters/IDPath"
    get:
      operationId: getItem
      summary: Get an item
      description: |
        Returns a single item with full sanitized HTML content. 404 if
        the item does not exist or does not belong to the caller (the
        two cases are not distinguished).
      tags: [items]
      responses:
        "200":
          description: The item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/items/{id}/star:
    parameters:
      - $ref: "#/components/parameters/IDPath"
    put:
      operationId: starItem
      summary: Star an item
      description: Marks an item as starred. Idempotent.
      tags: [items]
      responses:
        "204":
          description: Star applied.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: unstarItem
      summary: Remove a star from an item
      description: Removes a star from an item. Idempotent.
      tags: [items]
      responses:
        "204":
          description: Star removed.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/items/mark-read:
    post:
      operationId: markItemsRead
      summary: Mark items as read in bulk
      description: |
        Marks items as read. With an empty body, marks all of the
        caller's items as read. With `feed_id` or `list_id` set, the
        scope is restricted accordingly. The two filters are mutually
        exclusive.
      tags: [items]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkReadRequest"
      responses:
        "200":
          description: Mark complete.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkReadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: rdr_pat
      description: |
        Bearer token (rdr Personal Access Token, format
        `rdr_pat_<64 hex chars>`) issued via the Settings page.
  schemas:
...
    User:
      type: object
      required: [id, username, created_at]
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
          description: Database id of the user.
        username:
          type: string
          description: Login name (unique).
        created_at:
          type: string
          format: date-time
          description: When the user account was created.
    Feed:
      type: object
      required:
        - id
        - url
        - title
        - created_at
        - item_count
        - unread_count
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        list_id:
          type: integer
          format: int64
          nullable: true
          description: List this feed belongs to (if any).
        url:
          type: string
          format: uri
          description: Feed (RSS/Atom) URL.
        title:
          type: string
          description: |
            Feed title. May be empty until the first successful fetch
            populates it.
        site_url:
          type: string
          format: uri
          description: Canonical website for the feed (if known).
        favicon_url:
          type: string
          format: uri
          description: Favicon URL discovered for the feed (if known).
        last_fetched_at:
          type: string
          format: date-time
          nullable: true
          description: When the feed was last successfully fetched.
        last_fetch_error:
          type: string
          description: Most recent fetch error, if any.
        consecutive_failures:
          type: integer
          minimum: 0
          description: |
            Number of consecutive fetch failures since the last success.
            A non-zero value indicates the feed is currently broken.
        created_at:
          type: string
          format: date-time
        item_count:
          type: integer
          minimum: 0
          description: Total number of items stored for this feed.
        unread_count:
          type: integer
          minimum: 0
          description: Unread items in this feed for the caller.
    AddFeedRequest:
      type: object
      required: [url]
      additionalProperties: false
      properties:
        url:
          type: string
          format: uri
          description: |
            URL to add. May be a feed URL or a website URL — in the
            latter case the server attempts auto-discovery of the
            feed link from `<link rel="alternate">` tags.
...
  responses:
    BadRequest:
      description: The request is malformed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: |
        The Authorization header is missing, malformed, references an
        unknown token, or the token has expired. The same status is
        returned for every failure mode so unauthenticated callers
        cannot enumerate valid token identifiers.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: |
        The requested resource does not exist or does not belong to the
        caller. The same status is returned for both so callers cannot
        enumerate ids belonging to other users.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: |
        The request conflicts with current state — for example, an
        attempt to add a feed URL that the caller has already
        subscribed to.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    UnprocessableEntity:
      description: |
        The request body is well-formed JSON but is semantically
        invalid — for example, the supplied URL could not be reached
        or did not contain a discoverable feed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  parameters:
...
    IDPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
      description: Numeric id of the resource.
...