# internal/api/server.gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package api

import (
	"context"
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Defines values for HealthStatusStatus.
const (
	Ok HealthStatusStatus = "ok"
)

// Valid indicates whether the value is a known member of the HealthStatusStatus enum.
func (e HealthStatusStatus) Valid() bool {
	switch e {
	case Ok:
		return true
	default:
		return false
	}
}

// AddFeedRequest defines model for AddFeedRequest.
type AddFeedRequest struct {
	// Url URL to add. May be a feed URL or a website URL — in the
	// latter case the server attempts auto-discovery of the
	// feed link from `<link rel="alternate">` tags.
	Url string `json:"url"`
}

// Feed defines model for Feed.
type Feed struct {
	// ConsecutiveFailures Number of consecutive fetch failures since the last success.
	// A non-zero value indicates the feed is currently broken.
	ConsecutiveFailures *int      `json:"consecutive_failures,omitempty"`
	CreatedAt           time.Time `json:"created_at"`

	// FaviconUrl Favicon URL discovered for the feed (if known).
	FaviconUrl *string `json:"favicon_url,omitempty"`
	Id         int64   `json:"id"`

	// ItemCount Total number of items stored for this feed.
	ItemCount int `json:"item_count"`

	// LastFetchError Most recent fetch error, if any.
	LastFetchError *string `json:"last_fetch_error,omitempty"`

	// LastFetchedAt When the feed was last successfully fetched.
	LastFetchedAt *time.Time `json:"last_fetched_at,omitempty"`

	// ListId List this feed belongs to (if any).
	ListId *int64 `json:"list_id,omitempty"`

	// SiteUrl Canonical website for the feed (if known).
	SiteUrl *string `json:"site_url,omitempty"`

	// Title Feed title. May be empty until the first successful fetch
	// populates it.
	Title string `json:"title"`

	// UnreadCount Unread items in this feed for the caller.
	UnreadCount int `json:"unread_count"`

	// Url Feed (RSS/Atom) URL.
	Url string `json:"url"`
}

// HealthStatus defines model for HealthStatus.
type HealthStatus struct {
	// Status Always "ok". Reserved for future expansion.
	Status HealthStatusStatus `json:"status"`
}

// HealthStatusStatus Always "ok". Reserved for future expansion.
type HealthStatusStatus string

// Item defines model for Item.
type Item struct {
	// Content Full sanitized HTML body. Omitted from list responses.
	Content *string `json:"content,omitempty"`

	// Description Short summary or excerpt (sanitized).
	Description *string    `json:"description,omitempty"`
	FeedId      int64      `json:"feed_id"`
	FeedSiteUrl *string    `json:"feed_site_url,omitempty"`
	FeedTitle   string     `json:"feed_title"`
	FeedUrl     *string    `json:"feed_url,omitempty"`
	Id          int64      `json:"id"`
	PublishedAt *time.Time `json:"published_at,omitempty"`
	Read        bool       `json:"read"`
	Starred     bool       `json:"starred"`
	Title       string     `json:"title"`
	Url         string     `json:"url"`
}

// MarkReadRequest defines model for MarkReadRequest.
type MarkReadRequest struct {
	// FeedId Restrict the mark-read to items in this feed.
	FeedId *int64 `json:"feed_id,omitempty"`

	// ListId Restrict the mark-read to items in feeds belonging to this list.
	ListId *int64 `json:"list_id,omitempty"`
}

// MarkReadResponse defines model for MarkReadResponse.
type MarkReadResponse struct {
	// Marked Number of items that transitioned from unread to read.
	Marked int `json:"marked"`
}

// Problem RFC 7807 Problem Details. Returned for any 4xx or 5xx response.
type Problem struct {
	// Detail Human-readable explanation specific to this occurrence.
	Detail *string `json:"detail,omitempty"`

	// Instance URI reference identifying the specific occurrence.
	Instance *string `json:"instance,omitempty"`

	// Status HTTP status code.
	Status int `json:"status"`

	// Title Short, human-readable summary of the problem.
	Title string `json:"title"`

	// Type URI identifying the problem type. `about:blank` is used for
	// problems that don't have a dedicated type URI.
	Type string `json:"type"`
}

// SyncStatus defines model for SyncStatus.
type SyncStatus struct {
	// Syncing True if a feed sync is currently in progress.
	Syncing bool `json:"syncing"`
}

// User defines model for User.
type User struct {
	// CreatedAt When the user account was created.
	CreatedAt time.Time `json:"created_at"`

	// Id Database id of the user.
	Id int64 `json:"id"`

	// Username Login name (unique).
	Username string `json:"username"`
}

// IDPath defines model for IDPath.
type IDPath = int64

// PageParam defines model for PageParam.
type PageParam = int

// BadRequest RFC 7807 Problem Details. Returned for any 4xx or 5xx response.
type BadRequest = Problem

// Conflict RFC 7807 Problem Details. Returned for any 4xx or 5xx response.
type Conflict = Problem

// InternalServerError RFC 7807 Problem Details. Returned for any 4xx or 5xx response.
type InternalServerError = Problem

// NotFound RFC 7807 Problem Details. Returned for any 4xx or 5xx response.
type NotFound = Problem

// Unauthorized RFC 7807 Problem Details. Returned for any 4xx or 5xx response.
type Unauthorized = Problem

// UnprocessableEntity RFC 7807 Problem Details. Returned for any 4xx or 5xx response.
type UnprocessableEntity = Problem

// ListItemsParams defines parameters for ListItems.
type ListItemsParams struct {
	// Page 1-based page index. Page size is 50.
	Page *PageParam `form:"page,omitempty" json:"page,omitempty"`

	// FeedId Restrict to a single feed (must belong to the caller).
	FeedId *int64 `form:"feed_id,omitempty" json:"feed_id,omitempty"`

	// ListId Restrict to feeds in a specific list.
	ListId *int64 `form:"list_id,omitempty" json:"list_id,omitempty"`

	// Unread If true, only return unread items.
	Unread *bool `form:"unread,omitempty" json:"unread,omitempty"`

	// Starred If true, only return starred items.
	Starred *bool `form:"starred,omitempty" json:"starred,omitempty"`
}

// AddFeedJSONRequestBody defines body for AddFeed for application/json ContentType.
type AddFeedJSONRequestBody = AddFeedRequest

// MarkItemsReadJSONRequestBody defines body for MarkItemsRead for application/json ContentType.
type MarkItemsReadJSONRequestBody = MarkReadRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List feeds
	// (GET /api/v1/feeds)
	ListFeeds(w http.ResponseWriter, r *http.Request)
	// Add a feed
	// (POST /api/v1/feeds)
	AddFeed(w http.ResponseWriter, r *http.Request)
	// Trigger a feed sync
	// (POST /api/v1/feeds/sync)
	SyncFeeds(w http.ResponseWriter, r *http.Request)
	// Get sync status
	// (GET /api/v1/feeds/sync/status)
	GetSyncStatus(w http.ResponseWriter, r *http.Request)
	// Delete a feed
	// (DELETE /api/v1/feeds/{id})
	DeleteFeed(w http.ResponseWriter, r *http.Request, id IDPath)
	// Get a feed
	// (GET /api/v1/feeds/{id})
	GetFeed(w http.ResponseWriter, r *http.Request, id IDPath)
	// Liveness probe
	// (GET /api/v1/healthz)
	GetHealthz(w http.ResponseWriter, r *http.Request)
	// List items
	// (GET /api/v1/items)
	ListItems(w http.ResponseWriter, r *http.Request, params ListItemsParams)
	// Mark items as read in bulk
	// (POST /api/v1/items/mark-read)
	MarkItemsRead(w http.ResponseWriter, r *http.Request)
	// Get an item
	// (GET /api/v1/items/{id})
	GetItem(w http.ResponseWriter, r *http.Request, id IDPath)
	// Remove a star from an item
	// (DELETE /api/v1/items/{id}/star)
	UnstarItem(w http.ResponseWriter, r *http.Request, id IDPath)
	// Star an item
	// (PUT /api/v1/items/{id}/star)
	StarItem(w http.ResponseWriter, r *http.Request, id IDPath)
	// Identify the authenticated user
	// (GET /api/v1/me)
	GetMe(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ListFeeds operation middleware
func (siw *ServerInterfaceWrapper) ListFeeds(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListFeeds(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// AddFeed operation middleware
func (siw *ServerInterfaceWrapper) AddFeed(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AddFeed(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// SyncFeeds operation middleware
func (siw *ServerInterfaceWrapper) SyncFeeds(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SyncFeeds(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetSyncStatus operation middleware
func (siw *ServerInterfaceWrapper) GetSyncStatus(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetSyncStatus(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteFeed operation middleware
func (siw *ServerInterfaceWrapper) DeleteFeed(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id IDPath

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "int64"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteFeed(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetFeed operation middleware
func (siw *ServerInterfaceWrapper) GetFeed(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id IDPath

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "int64"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetFeed(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetHealthz operation middleware
func (siw *ServerInterfaceWrapper) GetHealthz(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealthz(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListItems operation middleware
func (siw *ServerInterfaceWrapper) ListItems(w http.ResponseWriter, r *http.Request) {

	var err error

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	// Parameter object where we will unmarshal all parameters from the context
	var params ListItemsParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "page", r.URL.Query(), &params.Page, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page", Err: err})
		return
	}

	// ------------- Optional query parameter "feed_id" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "feed_id", r.URL.Query(), &params.FeedId, runtime.BindQueryParameterOptions{Type: "integer", Format: "int64"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "feed_id", Err: err})
		return
	}

	// ------------- Optional query parameter "list_id" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "list_id", r.URL.Query(), &params.ListId, runtime.BindQueryParameterOptions{Type: "integer", Format: "int64"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "list_id", Err: err})
		return
	}

	// ------------- Optional query parameter "unread" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "unread", r.URL.Query(), &params.Unread, runtime.BindQueryParameterOptions{Type: "boolean", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "unread", Err: err})
		return
	}

	// ------------- Optional query parameter "starred" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "starred", r.URL.Query(), &params.Starred, runtime.BindQueryParameterOptions{Type: "boolean", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "starred", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListItems(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// MarkItemsRead operation middleware
func (siw *ServerInterfaceWrapper) MarkItemsRead(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.MarkItemsRead(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetItem operation middleware
func (siw *ServerInterfaceWrapper) GetItem(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id IDPath

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "int64"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetItem(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UnstarItem operation middleware
func (siw *ServerInterfaceWrapper) UnstarItem(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id IDPath

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "int64"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UnstarItem(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// StarItem operation middleware
func (siw *ServerInterfaceWrapper) StarItem(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id IDPath

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "int64"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.StarItem(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetMe operation middleware
func (siw *ServerInterfaceWrapper) GetMe(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetMe(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("GET "+options.BaseURL+"/api/v1/feeds", wrapper.ListFeeds)
	m.HandleFunc("POST "+options.BaseURL+"/api/v1/feeds", wrapper.AddFeed)
	m.HandleFunc("POST "+options.BaseURL+"/api/v1/feeds/sync", wrapper.SyncFeeds)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/feeds/sync/status", wrapper.GetSyncStatus)
	m.HandleFunc("DELETE "+options.BaseURL+"/api/v1/feeds/{id}", wrapper.DeleteFeed)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/feeds/{id}", wrapper.GetFeed)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/healthz", wrapper.GetHealthz)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/items", wrapper.ListItems)
	m.HandleFunc("POST "+options.BaseURL+"/api/v1/items/mark-read", wrapper.MarkItemsRead)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/items/{id}", wrapper.GetItem)
	m.HandleFunc("DELETE "+options.BaseURL+"/api/v1/items/{id}/star", wrapper.UnstarItem)
	m.HandleFunc("PUT "+options.BaseURL+"/api/v1/items/{id}/star", wrapper.StarItem)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/me", wrapper.GetMe)

	return m
}